+++ src/backend/core/tasks/item.py
@@ -51,6 +51,11 @@ def process_item_deletion(item_id):
 @app.task
 def rename_file(item_id, new_title):
     """Rename the file of an item. Update the filename and then rename the file on storage."""
+
+    if not new_title:
+        logger.error("New title is empty, skipping rename file")
+        return
+
     try:
         item = Item.objects.get(id=item_id)
     except Item.DoesNotExist: